bitwarden_backup/requirements.txt
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
cffi==1.17.1
cryptography==46.0.2
dateutils==0.6.12
pycparser==2.23
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
typing_extensions==4.13.2
